Implement conditional test skip based on pip version
