+++ tox.ini
@@ -11,6 +11,7 @@ deps =
     pytest-mypy
     pytest-timeout
     pytest-venv
+    packaging
     toml: toml
     pytoml: pytoml
 setenv =
